extend test to cover exceptions
